libcoap-bench/requirements.txt
# Data processing and analysis
pandas>=1.5.3
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0

# System utilities
python-dateutil>=2.8.2

# For parallel execution (if needed)
joblib>=1.2.0

# CSV handling
pyarrow>=11.0.0  # For faster CSV operations with pandas

# Serial communication
pyserial>=3.5

# USB communication
pyusb>=1.3.1
crc>=7.1.0
